Note Interface work
Created a note button that will open a "write note" interface. Can only
save 3 notes before the note button is deactivated. The "write note"
interface can be written in and the text saved as a datapacket, or it
can be closed and the text discarded. When discarded the note count
isn't updated. While the note interface is open all action in paused in
the game.